Added new status code and moved router handlers
