+++ src/config.ts
@@ -1,7 +1,5 @@
 import * as dotenv from 'dotenv';
-dotenv.config({
-  path: '${__dirname}/../.env',
-});
+dotenv.config();
 
 export const config = {
   db: {
