+++ src/models/status-codes.model.ts
@@ -2,4 +2,5 @@ export enum StatusCode {
   Success = 200,
   Invalid = 400,
   ExistingValue = 409,
+  UnprocessableEntity = 422,
 }
