+++ src/routers/exchanger.router.ts
@@ -1,13 +1,6 @@
-import { Request, Response, Router } from 'express';
-import { ExchangerService } from '../services/exhanger.service';
-import { StatusCode } from '../models/status-codes.model';
+import { Router } from 'express';
+import { ExchangerController } from '../controllers/exchanger.controller';
 
 export const exchangerRouter = Router();
-
-exchangerRouter.get('/', async (request: Request, response: Response) => {
-  const currentRate = await ExchangerService.getCurrentRate();
-  if (!currentRate) {
-    response.status(StatusCode.Invalid).json('Invalid status value');
-  }
-  response.status(StatusCode.Success).json(currentRate);
-});
+const exchangerController = new ExchangerController();
+exchangerRouter.get('/', exchangerController.getCurrentRate);
