+++ src/rules/subscription.rules.ts
@@ -1,0 +1,3 @@
+import { body } from 'express-validator';
+
+export const createSubscriprionRules = [body('email').trim().isEmail().withMessage('Specify valid email')];
